+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format of this changelog is based on
 
   - Added `SchematicDrivenLayout.filter_parameters` for sharing parameters between composite components and subcomponents
   - Added component style guide to docs
+  - Fixed bug where `map_metadata!` would map multiply-referenced structures multiple times
 
 ## 1.9.0 (2026-02-09)